// Da migliorare
1 . Tutta la procedura di log, per registrare gli errori o anomalie nella formatatura del file
2.migliorare l'algoratimo che tira fuori gli ingredienti , se provate non differenzia bene alcune stringhe e salva come la descrizione 1 melone ad esempio ma dovrebbe essere melone e 1 nel campo quantita ,uguale per i pesi in G / KG
3. Assolutamente trovare un modo per avere tutti i tipi piatti presenti nel e agganciarli alle ricette
// da iniziare

1.Fare qualche servizio per iniziare
2. iniziare a lavorare sul frontend

Abbiamo già un paio di entita e una bozza di database, dire che si puo iniziare a lavorare

+++ v1/src/main/java/com/DunnoWhatToEat/v1/Entity/Ingrediente.java
@@ -16,9 +16,9 @@ public class Ingrediente {
     private String nome;
 
     @Column
-    private int quantita;
+    private double quantita;
 
-    public void setQuantita(int quantita) {
+    public void setQuantita(double quantita) {
         this.quantita = quantita;
     }
 
@@ -26,7 +26,7 @@ public class Ingrediente {
         this.unita_misura = unita_misura;
     }
 
-    public int getQuantita() {
+    public double getQuantita() {
         return quantita;
     }
